requirements.txt
Flask
pandas
python-dotenv
openai
requests
beautifulsoup4
